+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/UTxO.hs
@@ -22,6 +22,7 @@ import Lens.Micro
 
 instance EraUTxO AllegraEra where
   type ScriptsNeeded AllegraEra = ShelleyScriptsNeeded AllegraEra
+  type StAnnTxCache AllegraEra = ()
 
   getConsumedValue pp lookupKeyDeposit = getConsumedCoin pp lookupKeyDeposit
 
@@ -36,3 +37,5 @@ instance EraUTxO AllegraEra where
   getWitsVKeyNeeded = getShelleyWitsVKeyNeeded
 
   getMinFeeTxUtxo pp tx _ = getShelleyMinFeeTxUtxo pp tx
+
+  getCacheStAnnTx _ = mempty