+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Rules/Ledgers.hs
@@ -23,10 +23,9 @@ import Cardano.Ledger.Alonzo.Rules.Utxow (AlonzoUtxowPredFailure)
 import Cardano.Ledger.BaseTypes (ShelleyBase, epochInfo, systemStart)
 import Cardano.Ledger.Core
 import Cardano.Ledger.Shelley.API.Mempool (ApplyTx (..))
-import Cardano.Ledger.Shelley.Core (EraGov)
-import Cardano.Ledger.Shelley.LedgerState (LedgerState (..), UTxOState (..))
+import Cardano.Ledger.Shelley.LedgerState (LedgerState (..))
 import qualified Cardano.Ledger.Shelley.Rules as Shelley
-import Cardano.Ledger.State (CertState, EraStake)
+import Cardano.Ledger.State
 import Cardano.Slotting.EpochInfo.Extend (unsafeLinearExtendEpochInfo)
 import Control.Monad (foldM)
 import Control.Monad.Trans.Reader (asks)
@@ -42,6 +41,7 @@ import Control.State.Transition (
 import Data.Default (Default)
 import Data.Foldable (toList)
 import Data.Sequence (Seq)
+import Lens.Micro ((^.))
 
 type instance EraRuleFailure "LEDGERS" AlonzoEra = Shelley.ShelleyLedgersPredFailure AlonzoEra
 
@@ -118,19 +118,24 @@ alonzoLedgersTransition ::
   ) =>
   TransitionRule (LEDGERS era)
 alonzoLedgersTransition = do
-  TRC (Shelley.LedgersEnv slot epochNo pp account, ls, txs) <- judgmentContext
+  TRC (Shelley.LedgersEnv slot epochNo pp account, initLedgerState, txs) <- judgmentContext
   ei <- liftSTS $ asks epochInfo
   sysStart <- liftSTS $ asks systemStart
-  foldM
-    ( \ !ls' (ix, tx) ->
-        let utxo = utxosUtxo (lsUTxOState ls')
-            stAnnTx = mkStAnnTx (unsafeLinearExtendEpochInfo slot ei) sysStart pp utxo tx
-         in trans @(EraRule "LEDGER" era) $
-              TRC (Shelley.LedgerEnv slot (Just epochNo) ix pp account, ls', stAnnTx)
-    )
-    ls
-    $ zip [minBound ..]
-    $ toList txs
+  (finalLedgerState, _) <-
+    foldM -- cache is lazy on purpose, since it is unused when validation is off
+      ( \(!curLedgerState, stAnnTxCache) (ix, tx) -> do
+          let utxo = curLedgerState ^. utxoL
+              stAnnTx =
+                mkStAnnTx (unsafeLinearExtendEpochInfo slot ei) sysStart pp utxo stAnnTxCache tx
+          newLedgerState <-
+            trans @(EraRule "LEDGER" era) $
+              TRC (Shelley.LedgerEnv slot (Just epochNo) ix pp account, curLedgerState, stAnnTx)
+          pure (newLedgerState, getCacheStAnnTx stAnnTx)
+      )
+      (initLedgerState, mempty)
+      $ zip [minBound ..]
+      $ toList txs
+  pure finalLedgerState
 
 instance
   ( Era era