+++ eras/alonzo/impl/testlib/Test/Cardano/Ledger/Alonzo/TreeDiff.hs
@@ -153,16 +153,20 @@ instance
   ) =>
   ToExpr (AlonzoStAnnTx TopTx era)
   where
-  toExpr stAnnTx@(AlonzoStAnnTx _ _ _ _ _) =
+  toExpr stAnnTx@(AlonzoStAnnTx _ _ _ _ _ _) =
     let AlonzoStAnnTx {..} = stAnnTx
      in Rec "AlonzoStAnnTx" $
           OMap.fromList
             [ ("asatTx", toExpr asatTx)
             , ("asatScriptsNeeded", toExpr asatScriptsNeeded)
             , ("asatScriptsProvided", toExpr asatScriptsProvided)
+            , ("asatPlutusRunnableCache", toExpr asatPlutusRunnableCache)
             , ("asatPlutusLanguagesUsed", toExpr asatPlutusLanguagesUsed)
             , ("asatPlutusScriptsWithContext", toExpr asatPlutusScriptsWithContext)
             ]
+
+instance ToExpr (SupportedPlutusRunnable era) where
+  toExpr (SupportedPlutusRunnable spr) = toExpr spr
 
 -- Plutus/TxInfo
 instance ToExpr (AlonzoContextError era)