+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/Rules/Ledgers.hs
@@ -19,12 +19,12 @@ import Cardano.Ledger.Babbage.Era (BabbageEra, LEDGER, LEDGERS)
 import Cardano.Ledger.Babbage.Rules.Ledger ()
 import Cardano.Ledger.Babbage.Rules.Utxo (BabbageUtxoPredFailure)
 import Cardano.Ledger.Babbage.Rules.Utxow (BabbageUtxowPredFailure)
-import Cardano.Ledger.Babbage.State
 import Cardano.Ledger.BaseTypes (ShelleyBase, epochInfo, systemStart)
 import Cardano.Ledger.Core
 import Cardano.Ledger.Shelley.API.Mempool (ApplyTx (..))
-import Cardano.Ledger.Shelley.LedgerState (LedgerState (..), UTxOState (..))
+import Cardano.Ledger.Shelley.LedgerState (LedgerState (..))
 import qualified Cardano.Ledger.Shelley.Rules as Shelley
+import Cardano.Ledger.State
 import Control.Monad (foldM)
 import Control.Monad.Trans.Reader (asks)
 import Control.State.Transition (
@@ -39,6 +39,7 @@ import Control.State.Transition (
 import Data.Default (Default)
 import Data.Foldable (toList)
 import Data.Sequence (Seq)
+import Lens.Micro ((^.))
 
 type instance EraRuleFailure "LEDGERS" BabbageEra = Shelley.ShelleyLedgersPredFailure BabbageEra
 
@@ -121,19 +122,24 @@ ledgersTransition ::
   ) =>
   TransitionRule (LEDGERS era)
 ledgersTransition = do
-  TRC (Shelley.LedgersEnv slot epochNo pp account, ls, txs) <- judgmentContext
+  TRC (Shelley.LedgersEnv slot epochNo pp account, initLedgerState, txs) <- judgmentContext
   ei <- liftSTS $ asks epochInfo
   sysStart <- liftSTS $ asks systemStart
-  foldM
-    ( \ !ls' (ix, tx) ->
-        let utxo = utxosUtxo (lsUTxOState ls')
-            stAnnTx = mkStAnnTx ei sysStart pp utxo tx
-         in trans @(EraRule "LEDGER" era) $
-              TRC (Shelley.LedgerEnv slot (Just epochNo) ix pp account, ls', stAnnTx)
-    )
-    ls
-    $ zip [minBound ..]
-    $ toList txs
+  (finalLedgerState, _) <-
+    foldM -- cache is lazy on purpose, since it is unused when validation is off
+      ( \(!curLedgerState, stAnnTxCache) (ix, tx) -> do
+          let
+            utxo = curLedgerState ^. utxoL
+            stAnnTx = mkStAnnTx ei sysStart pp utxo stAnnTxCache tx
+          newLedgerState <-
+            trans @(EraRule "LEDGER" era) $
+              TRC (Shelley.LedgerEnv slot (Just epochNo) ix pp account, curLedgerState, stAnnTx)
+          pure (newLedgerState, getCacheStAnnTx stAnnTx)
+      )
+      (initLedgerState, mempty)
+      $ zip [minBound ..]
+      $ toList txs
+  pure finalLedgerState
 
 instance
   ( Era era