+++ eras/mary/impl/src/Cardano/Ledger/Mary.hs
@@ -46,7 +46,7 @@ instance ApplyTx MaryEra where
     deriving (Eq, Show)
     deriving newtype (EncCBOR, DecCBOR, Semigroup, Generic)
 
-  mkStAnnTx _ _ _ _ = id
+  mkStAnnTx _ _ _ _ _ = id
 
   internalApplyTxWithValidation = defaultApplyTxWithValidation @"LEDGER" MaryApplyTxError