+++ eras/mary/impl/src/Cardano/Ledger/Mary/UTxO.hs
@@ -38,6 +38,7 @@ import Lens.Micro
 
 instance EraUTxO MaryEra where
   type ScriptsNeeded MaryEra = ShelleyScriptsNeeded MaryEra
+  type StAnnTxCache MaryEra = ()
 
   getConsumedValue = getConsumedMaryValue
 
@@ -52,6 +53,8 @@ instance EraUTxO MaryEra where
   getWitsVKeyNeeded = getShelleyWitsVKeyNeeded
 
   getMinFeeTxUtxo pp tx _ = getShelleyMinFeeTxUtxo pp tx
+
+  getCacheStAnnTx _ = mempty
 
 -- | Calculate the value consumed by the transation.
 --