+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/MiniTrace.hs
@@ -583,6 +583,7 @@ constrainedUtxo =
             (systemStart testGlobals)
             (uePParams uecUtxoEnv)
             uecUTxO
+            mempty
             tx
     )