+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/Properties.hs
@@ -120,6 +120,7 @@ genTxAndLEDGERState sizes = do
                 (systemStart testGlobals)
                 pp
                 (utxosUtxo (lsUTxOState ledgerState))
+                mempty
                 tx
         pure $ TRC (ledgerEnv, ledgerState, stAnnTx)
   (trc, genstate) <- runGenRS sizes (initStableFields >> genT)